+++ fixtures/ext-types/proc-macro-lib/tests/bindings/test_imported_types.py
@@ -35,7 +35,7 @@ class TestIt(unittest.TestCase):
         self.assertEqual(get_maybe_urls([url, None]), [url, None])
 
     def test_get_uniffi_one_type(self):
-        t1 = UniffiOneType("hello")
+        t1 = UniffiOneType(sval="hello")
         self.assertEqual(t1, get_uniffi_one_type(t1))
         self.assertEqual(t1, get_maybe_uniffi_one_type(t1))
         self.assertEqual(None, get_maybe_uniffi_one_type(None))
@@ -44,7 +44,7 @@ class TestIt(unittest.TestCase):
 
     def test_async(self):
         async def test():
-            t1 = UniffiOneType("hello")
+            t1 = UniffiOneType(sval="hello")
             # This async function comes from the `uniffi-one` crate
             self.assertEqual(await get_uniffi_one_async(), UniffiOneEnum.ONE)
             # This async function comes from the `proc-macro-lib` crate
@@ -52,7 +52,7 @@ class TestIt(unittest.TestCase):
         asyncio.run(test())
 
     def test_get_uniffi_one_proc_macro_type(self):
-        t1 = UniffiOneProcMacroType("hello")
+        t1 = UniffiOneProcMacroType(sval="hello")
         self.assertEqual(t1, get_uniffi_one_proc_macro_type(t1))
 
     def test_get_uniffi_one_enum(self):