+++ fixtures/struct-default-values/tests/test_generated_bindings.rs
@@ -1,0 +1,5 @@
+uniffi::build_foreign_language_testcases!(
+    "tests/bindings/test_struct_default_values.py",
+    "tests/bindings/test_struct_default_values.kts",
+    "tests/bindings/test_struct_default_values.swift",
+);